Refactor source layout into modular directories
