+++ src/lexer/lexer.c
@@ -1,4 +1,4 @@
-#include "dream.h"
+#include "lexer.h"
 #include <ctype.h>
 #include <stdlib.h>
 #include <string.h>
