+++ src/parser/expression.c
@@ -1,0 +1,40 @@
+#include "parser.h"
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+Node *parse_expression(Lexer *lexer, Token *token) {
+  int unary_minus = 0;
+  if (token->type == TOKEN_MINUS) {
+    unary_minus = 1;
+    free(token->value);
+    *token = next_token(lexer);
+  }
+  if (token->type != TOKEN_IDENTIFIER && token->type != TOKEN_NUMBER &&
+      token->type != TOKEN_STRING) {
+    fprintf(stderr, "Expected identifier, number or string\n");
+    exit(1);
+  }
+  NodeType left_type = NODE_IDENTIFIER;
+  if (token->type == TOKEN_NUMBER)
+    left_type = NODE_NUMBER;
+  else if (token->type == TOKEN_STRING)
+    left_type = NODE_STRING;
+  Node *left = create_node(left_type, token->value, NULL, NULL, NULL);
+  *token = next_token(lexer);
+  if (unary_minus)
+    left = create_node(NODE_UNARY_OP, "-", left, NULL, NULL);
+  if ((token->type == TOKEN_PLUS || token->type == TOKEN_MINUS ||
+       token->type == TOKEN_STAR || token->type == TOKEN_SLASH ||
+       token->type == TOKEN_PERCENT || token->type == TOKEN_LT ||
+       token->type == TOKEN_GT || token->type == TOKEN_LE ||
+       token->type == TOKEN_GE || token->type == TOKEN_EQEQ ||
+       token->type == TOKEN_NEQ) &&
+      left_type != NODE_STRING) {
+    char *op = token->value;
+    *token = next_token(lexer);
+    Node *right = parse_expression(lexer, token);
+    return create_node(NODE_BINARY_OP, op, left, right, NULL);
+  }
+  return left;
+}
